Added a "TakesPlayerInput" base class, updated gunmodels
We added a "TakesPlayerInput" base class intended for classes that use player input to inherit from. This base class makes it simple to disable specific player inputs while in the pause menu. The new gun models (Sniper, Rifle) are made from Unity primitive objects.

+++ fps-1/Assets/Scripts/Weapon.cs
@@ -6,8 +6,9 @@ public class Weapon {
     public string name = "Pistol";
     public int damage = 10;
     public float range = 200f;
-    public int fireRate = 0;
+    public float fireRate = 0f;
     public GameObject weaponModel = null;
+
 
     void Awake()
     {
